types are now inferred rather than defined through interfaces. Changed names of files and variables to make terminology clearer and code neater. Took out unnecessary files and code.
