+++ src/pages/book/[bookid]/[chapterid]/addSection.tsx
@@ -1,36 +1,17 @@
 "use client";
-import { signIn, signOut, useSession } from "next-auth/react";
 import Head from "next/head";
-import Link from "next/link";
-import { useState, useEffect } from "react";
 import { useRouter } from "next/router";
-import { api } from "~/utils/api";
-import { z } from "zod";
-import { useForm, useFieldArray, Controller, Control } from "react-hook-form";
-import { Button } from "@/components/ui/button";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-import { zodResolver } from "@hookform/resolvers/zod";
-import BookForm from "~/components/BookForm";
-import BookRender from "~/components/BookRender";
-import AutoBookForm from "~/components/AutoBookForm";
-import AutoChapterForm from "~/components/AutoChapterForm";
-import AutoSectionForm from "~/components/AutoSectionForm";
+import BookRenderer from "~/components/BookRenderer";
+
+import SectionForm from "~/components/SectionForm";
 export default function AddSectionPage() {
   const router = useRouter();
   const { bookid, chapterid } = router.query; // `bookid` matches the dynamic segment name
- 
+
   return (
     <>
       <Head>
-        <title>Create T3 App</title>
+        <title>Add Section</title>
         <meta name="description" content="Generated by create-t3-app" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
@@ -40,10 +21,13 @@ export default function AddSectionPage() {
         </div>
         <div className="flex w-full">
           <div className="w-1/2">
-            <AutoSectionForm bookId={bookid as string} chapterId={chapterid as string} />
+            <SectionForm
+              bookId={bookid as string}
+              chapterId={chapterid as string}
+            />
           </div>
           <div className="w-1/2">
-            <BookRender />
+            <BookRenderer />
           </div>
         </div>
       </main>
