+++ src/pages/book/[bookid]/addChapter.tsx
@@ -1,27 +1,8 @@
 "use client";
-import { signIn, signOut, useSession } from "next-auth/react";
 import Head from "next/head";
-import Link from "next/link";
-import { useState, useEffect } from "react";
 import { useRouter } from "next/router";
-import { api } from "~/utils/api";
-import { z } from "zod";
-import { useForm, useFieldArray, Controller, Control } from "react-hook-form";
-import { Button } from "@/components/ui/button";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-import { zodResolver } from "@hookform/resolvers/zod";
-import BookForm from "~/components/BookForm";
-import BookRender from "~/components/BookRender";
-import AutoBookForm from "~/components/AutoBookForm";
-import AutoChapterForm from "~/components/AutoChapterForm";
+import BookRender from "~/components/BookRenderer";
+import AutoChapterForm from "~/components/ChapterForm";
 
 export default function AddChapterPage() {
   const router = useRouter();
@@ -29,7 +10,7 @@ export default function AddChapterPage() {
   return (
     <>
       <Head>
-        <title>Create T3 App</title>
+        <title>Add Chapter</title>
         <meta name="description" content="Generated by create-t3-app" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
