+++ src/pages/index.tsx
@@ -1,34 +1,15 @@
 "use client";
 import { signIn, signOut, useSession } from "next-auth/react";
 import Head from "next/head";
-import Link from "next/link";
-
 import { api } from "~/utils/api";
-import { z } from "zod";
-import { useForm, useFieldArray, Controller, Control } from "react-hook-form";
-
-import { Button } from "@/components/ui/button";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-import { zodResolver } from "@hookform/resolvers/zod";
+import BookRenderer from "~/components/BookRenderer";
 import BookForm from "~/components/BookForm";
-import BookRender from "~/components/BookRender";
-import AutoBookForm from "~/components/AutoBookForm";
 
 export default function Home() {
-  const hello = api.post.hello.useQuery({ text: "from tRPC" });
-
   return (
     <>
       <Head>
-        <title>Create T3 App</title>
+        <title>BookWyrm</title>
         <meta name="description" content="Generated by create-t3-app" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
@@ -40,8 +21,8 @@ export default function Home() {
         </div>
 
         <div className="flex w-full">
-          <div className="w-1/2">{AutoBookForm()}</div>
-          <div className="w-1/2">{BookRender()}</div>
+          <div className="w-1/2">{BookForm()}</div>
+          <div className="w-1/2">{BookRenderer()}</div>
         </div>
       </main>
     </>
